fix(image): correct zoom-toward-cursor math and edge cases
- Guard against zero-dimension images in compute_fit_scale
- Fix keyboard zoom anchoring at (0,0) by using last mouse position
- Convert window coords to content-area-local coords for zoom math
- Fix u32 overflow in pixel index computation for large images
- Recompute fit scale on window resize for image mode
- Add test for zero-dimension image edge case

+++ src/image/render.rs
@@ -88,7 +88,7 @@ pub fn render_image(
                 && (img_x as u32) < img_w
                 && (img_y as u32) < img_h
             {
-                let src_idx = ((img_y as u32 * img_w + img_x as u32) * 4) as usize;
+                let src_idx = (img_y as usize * img_w as usize + img_x as usize) * 4;
 
                 if src_idx + 3 < image.pixels.len() {
                     let r = image.pixels[src_idx] as u32;
